refactor: move schema check to write request
